Intake spinners code
